fix: reload page after one day to avoid stale data
Our application is a SPA (Single Page Application).
In our app, we send requests to API webservices to get data
if that data is not in the redux store.
In other words, if data is already in the redux state, we won't
send requests.
There will be a problem of stale data.

For example,
1. end user goes to page A. App will fetch page A data and store it
in redux state.
2. end user goes to page B. App will fetch page B data and store it
in redux state.
3. end user goes to page A again. App won't fetch page A data since
it already in the redux state.

If end users don't refresh the web page, the data will be always in
the redux store. Therefore, end users might see the stale data
even the data is already updated.