fix: handle other errors except for 404 and 500 errors
When error occurs, maybe redux action errors, maybe express errors, we
treat them only two ways.
One is 404 error, and the other is 500 error.

However, there are some errors, such as 406 Not Acceptable error, which
should not be redirected to 500 error page.

In this patch, we log information of error whose status code < 500, and
trigger error reportering if error's status code >= 500.