fix: change font loader
This patch changes font loader to address white page issue in mobile safari.

Before this change, we used adobe font embeded code for loading the web
fonts. However, it causes bizarre behavior which renders white page and
seems like the rendering is blocked by script executing.

That's the reason that we decided to change the font loader to webfont
loader, which is a tool to load webfont, and also change the webfont to
google font.