fix: unexpected DOMExceptionError due to rendering promote `Banner`
This bug is hard to reproduce.

Currently, we are not so sure what kind of environment will trigger
this error.

However, my computer is macOS and its os version is v14.5.

In this situation, when react renders `Banner` promote component,
rather than `Popup` component, the DOMExceptionError occurs.

This patch is simply add `<div>` outside `<PromoContext.Provider>` to
make sure that react will find a parent node where to replace `Popup`
component to `Banner` under.